sysloadavg: fix SEGFAULT caused by incorrect printf type
Commit 12e10224738fcf43919982ee3207d65ed18e2667 added support for
a new resource "sysloadavg". In http/cervlet.c line 1843 the call
to out_print in the format specifier defined a float or double ("%.1f")
be provided in the 4th argument. The original value provided was of type
int.

The fix is to explicitly typecast the 4th argument to type double.

Closes: https://tickets.cumulusnetworks.com/browse/CM-5345
Fixes: 12e10224738fcf43919982ee3207d65ed18e2667

+++ http/cervlet.c
@@ -1832,7 +1832,7 @@ static void print_service_rules_resource(HttpResponse res, Service_T s) {
         case RESOURCE_ID_SYSLOAD1:
         case RESOURCE_ID_SYSLOAD5:
         case RESOURCE_ID_SYSLOAD15:
-          out_print(res, "If %s %.1f %s ", operatornames[q->operator], q->limit, Util_getEventratio(a->failed, buf, sizeof(buf)));
+          out_print(res, "If %s %.1f %s ", operatornames[q->operator], (double) q->limit, Util_getEventratio(a->failed, buf, sizeof(buf)));
           out_print(res, "then %s ", Util_describeAction(a->failed, buf, sizeof(buf)));
           out_print(res, "else if succeeded %s ", Util_getEventratio(a->succeeded, buf, sizeof(buf)));
           out_print(res, "then %s", Util_describeAction(a->succeeded, buf, sizeof(buf)));
